Introduce the ability to edit note text inline.
Remove the ability to use Markdown in notes and quotations.

Notes and quotations are now displayed with the whitespace that the user
specified.

Simplify handling of the escape key.

Fix a bug that prevented the help window from being displayed.

Fix a bug where note modification times were displayed on the same line
as the note.