Improve drag hints so that hints about the subnotes are only shown if a dragged note has subnotes.
Improve drag hints so that hints about whether the subnotes are being
dragged are shown whether or note they are being dragged.

Improve drag hints so that pressing or unpressing Ctrl or Alt updates
the drag hint immediately.

Fix some bugs related to showing the + icon when moving a note left or
right, or dragging it. Most of these bugs were introduced since the past
release when the + icon was changed to not display by default, but some
were introduced earlier.

Slightly optimize the algorithm that shows or hides pluses up to the
root of a note tree by short circuiting it.