Update the shortcut keys so they can be used during inline editing. Change the c, Delete, m, i, u, b, a, +, -, Left, Right, Up, Down Ctrl+Left, Ctrl+Right, Ctrl+Up and Ctrl+Down shortcut key to require Alt as well. Remove the h, for help, shortcut key. Remove the a, for select all, shortcut key. Remove the Ctrl+space, for deselect all, shortcut key.
Fix an issue where FireFox would forward key events to handlers that
were not registered at the time of the event.

Make local changes to mousetrap.min.js so that Mousetrap.js can be
forced to use the key event specified by the application. This is
because on Firefox the key combo Alt+- is delivered with keypress rather
than keydown as Mousetrap.js expects.